# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tal'ah API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ReadinessStatus {
  status: string;
  database: string;
}

export interface VersionInfo {
  name: string;
  version: string;
  commit: string;
  nodeEnv: string;
}

export interface User {
  id: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  nickname?: string | null;
  /** @nullable */
  gender?: string | null;
  /** @nullable */
  city?: string | null;
  onboarded: boolean;
  createdAt: string;
}

export interface UserPatch {
  [key: string]: unknown;
}

export interface MeetupRequest {
  id?: string;
  userId?: string;
  status?: string;
  [key: string]: unknown;
}

export type CreateRequestMeetupType =
  (typeof CreateRequestMeetupType)[keyof typeof CreateRequestMeetupType];

export const CreateRequestMeetupType = {
  coffee: "coffee",
  dinner: "dinner",
} as const;

export type CreateRequestPreferredTime =
  (typeof CreateRequestPreferredTime)[keyof typeof CreateRequestPreferredTime];

export const CreateRequestPreferredTime = {
  morning: "morning",
  afternoon: "afternoon",
  evening: "evening",
} as const;

export interface CreateRequest {
  meetupType: CreateRequestMeetupType;
  preferredDate: string;
  preferredTime: CreateRequestPreferredTime;
  area: string;
  friendEmail?: string;
}

export type RequestInvitationStatus =
  (typeof RequestInvitationStatus)[keyof typeof RequestInvitationStatus];

export const RequestInvitationStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
  expired: "expired",
  finalized: "finalized",
} as const;

export interface RequestInvitation {
  id: string;
  requestId: string;
  requesterId: string;
  invitedEmail: string;
  /** @nullable */
  inviteeUserId?: string | null;
  status: RequestInvitationStatus;
  expiresAt: string;
  /** @nullable */
  respondedAt?: string | null;
  createdAt: string;
}

export interface Group {
  id?: string;
  status?: string;
  [key: string]: unknown;
}

export interface AdminAuditLog {
  id?: string;
  action?: string;
  targetTable?: string;
  /** @nullable */
  targetId?: string | null;
  createdAt?: string;
  [key: string]: unknown;
}

export interface PaginatedAdminAuditLog {
  data: AdminAuditLog[];
  total: number;
  hasMore: boolean;
}

export type LimitParameter = number;

export type OffsetParameter = number;

export type SendEmailLoginCodeBody = {
  email: string;
};

export type SendEmailLoginCode200 = {
  ok: boolean;
  /** Development only */
  code?: string;
};

export type VerifyEmailLoginCodeBody = {
  email: string;
  /**
   * @minLength 4
   * @maxLength 6
   */
  code: string;
};

export type VerifyEmailLoginCode200 = {
  token: string;
  user: User;
};

export type ClaimInvitationBody = {
  token: string;
};

export type RespondToInvitationBodyResponse =
  (typeof RespondToInvitationBodyResponse)[keyof typeof RespondToInvitationBodyResponse];

export const RespondToInvitationBodyResponse = {
  accepted: "accepted",
  declined: "declined",
} as const;

export type RespondToInvitationBody = {
  response: RespondToInvitationBodyResponse;
};

export type GetAdminAuditLogsParams = {
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: LimitParameter;
  /**
   * @minimum 0
   */
  offset?: OffsetParameter;
};