# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Tal'ah API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      security: [{ bearerAuth: [] }]
      summary: List current user's requests
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetupRequest"
    post:
      operationId: createRequest
      tags: [requests]
      security: [{ bearerAuth: [] }]
      summary: Create a meetup request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequest"
      responses:
        "201":
          description: Created request
...
  /invitations:
    get:
      operationId: getInvitations
      tags: [invitations]
      security: [{ bearerAuth: [] }]
      summary: List invitations for the current requester or recipient
      responses:
        "200":
          description: Invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RequestInvitation"
  /invitations/claim:
    post:
      operationId: claimInvitation
      tags: [invitations]
      security: [{ bearerAuth: [] }]
      summary: Claim an invitation linked from email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Claimed invitation
  /invitations/{id}/respond:
    post:
      operationId: respondToInvitation
      tags: [invitations]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response: { type: string, enum: [accepted, declined] }
      responses:
        "200":
          description: Updated invitation
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    MeetupRequest:
      type: object
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        status: { type: string }
    CreateRequest:
      type: object
      required: [meetupType, preferredDate, preferredTime, area]
      properties:
        meetupType: { type: string, enum: [coffee, dinner] }
        preferredDate: { type: string }
        preferredTime: { type: string, enum: [morning, afternoon, evening] }
        area: { type: string }
        friendEmail: { type: string, format: email }
    RequestInvitation:
      type: object
      required: [id, requestId, requesterId, invitedEmail, status, expiresAt, createdAt]
      properties:
        id: { type: string, format: uuid }
        requestId: { type: string, format: uuid }
        requesterId: { type: string, format: uuid }
        invitedEmail: { type: string, format: email }
        inviteeUserId: { type: [string, "null"], format: uuid }
        status: { type: string, enum: [pending, accepted, declined, expired, finalized] }
        expiresAt: { type: string, format: date-time }
        respondedAt: { type: [string, "null"], format: date-time }
        createdAt: { type: string, format: date-time }
...