# lib/api-zod/src/generated/types/createRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tal'ah API specification
 * OpenAPI spec version: 0.2.0
 */
import type { CreateRequestMeetupType } from "./createRequestMeetupType";
import type { CreateRequestPreferredTime } from "./createRequestPreferredTime";

export interface CreateRequest {
  meetupType: CreateRequestMeetupType;
  preferredDate: string;
  preferredTime: CreateRequestPreferredTime;
  area: string;
  friendEmail?: string;
}