# lib/api-zod/src/generated/types/meetupRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tal'ah API specification
 * OpenAPI spec version: 0.2.0
 */

export interface MeetupRequest {
  id?: string;
  userId?: string;
  status?: string;
  [key: string]: unknown;
}