# lib/api-zod/src/generated/types/requestInvitation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tal'ah API specification
 * OpenAPI spec version: 0.2.0
 */
import type { RequestInvitationStatus } from "./requestInvitationStatus";

export interface RequestInvitation {
  id: string;
  requestId: string;
  requesterId: string;
  invitedEmail: string;
  /** @nullable */
  inviteeUserId?: string | null;
  status: RequestInvitationStatus;
  expiresAt: Date;
  /** @nullable */
  respondedAt?: Date | null;
  createdAt: Date;
}